feat: deploy on sepolia, non-verified

+++ test/Mysterion.t.sol
@@ -1,21 +1,16 @@
 // SPDX-License-Identifier: MIT
 pragma solidity ^0.8.20;
 
-
 import "forge-std/Test.sol";
 import "../src/Mysterion.sol";
 
-
 contract TestERC721 is Test {
     Mysterion public erc721;
-    
-    event Transfer(address indexed from, address indexed to, uint indexed id);
-    event Approval(address indexed owner, address indexed spender, uint indexed id);
-    event ApprovalForAll(
-        address indexed owner,
-        address indexed operator,
-        bool approved
-    );
+
+    event Transfer(address indexed from, address indexed to, uint256 indexed id);
+    event Approval(address indexed owner, address indexed spender, uint256 indexed id);
+    event ApprovalForAll(address indexed owner, address indexed operator, bool approved);
+
     address public alice = address(1);
     address public bob = address(2);
     address public cat = address(3);
@@ -26,47 +21,47 @@ contract TestERC721 is Test {
 
     function testBaseURI() public {
         string memory uri = erc721.baseURI();
-        assertEq(uri,"https://api.mysterion.com/api/v1/token/");
+        assertEq(uri, "https://api.mysterion.com/api/v1/token/");
     }
 
     function testSafeMint() public {
         vm.prank(alice);
         erc721.safeMint(alice); // tokenId 0 mint
-        
+
         address zero_owner = erc721.ownerOf(0);
-        uint alice_balance = erc721.balanceOf(alice);
+        uint256 alice_balance = erc721.balanceOf(alice);
         assertEq(alice_balance, 1);
         assertEq(alice, zero_owner);
     }
 
     function testOnlyMinterSafeMint() public {
         vm.startPrank(bob);
-        
+
         vm.expectRevert(); // "Only minter can mint the token"
         erc721.safeMint(bob);
-        
+
         vm.stopPrank();
     }
 
     function testTokenURI() public {
         vm.prank(alice);
         erc721.safeMint(alice);
-        
+
         string memory uri = erc721.tokenURI(0);
-        assertEq(uri,"https://api.mysterion.com/api/v1/token/0");
+        assertEq(uri, "https://api.mysterion.com/api/v1/token/0");
 
         vm.prank(alice);
         erc721.safeMint(bob);
         string memory uri2 = erc721.tokenURI(1);
-        assertEq(uri2,"https://api.mysterion.com/api/v1/token/1");
+        assertEq(uri2, "https://api.mysterion.com/api/v1/token/1");
     }
 
     function testSetBaseURI() public {
         vm.prank(alice);
         erc721.setBaseURI("https://api.mysterion.com/api/v2/token/");
-        
+
         string memory uri = erc721.baseURI();
-        assertEq(uri,"https://api.mysterion.com/api/v2/token/");
+        assertEq(uri, "https://api.mysterion.com/api/v2/token/");
 
         vm.prank(alice);
         erc721.safeMint(alice);
@@ -83,7 +78,7 @@ contract TestERC721 is Test {
         erc721.safeMint(bob);
         erc721.safeMint(bob);
         vm.stopPrank();
-        uint bob_balance = erc721.balanceOf(bob);
+        uint256 bob_balance = erc721.balanceOf(bob);
         assertEq(bob_balance, 6);
     }
 
@@ -91,7 +86,7 @@ contract TestERC721 is Test {
         vm.prank(alice);
         erc721.safeMint(alice);
         address owner_token_0 = erc721.ownerOf(0);
-        uint alice_balance = erc721.balanceOf(alice);
+        uint256 alice_balance = erc721.balanceOf(alice);
         assertEq(alice, owner_token_0);
         assertEq(alice_balance, 1);
 
@@ -101,16 +96,15 @@ contract TestERC721 is Test {
         owner_token_0 = erc721.ownerOf(0);
         assertEq(bob, owner_token_0);
         alice_balance = erc721.balanceOf(alice);
-        uint bob_balance = erc721.balanceOf(bob);
+        uint256 bob_balance = erc721.balanceOf(bob);
         assertEq(alice_balance, 0);
         assertEq(bob_balance, 1);
-    }   
-
+    }
 
     function testApprove() public {
         vm.prank(alice);
         erc721.safeMint(bob); // token 0
-        
+
         vm.prank(bob);
         erc721.approve(alice, 0);
         address owner_token_0 = erc721.ownerOf(0);
@@ -125,28 +119,25 @@ contract TestERC721 is Test {
     function testSetApprovalForAll() public {
         vm.prank(alice);
         erc721.safeMint(bob);
-        
+
         vm.prank(bob);
         erc721.setApprovalForAll(alice, true);
         bool _approved = erc721.isApprovedForAll(bob, alice);
-        assertEq(_approved,true);
+        assertEq(_approved, true);
     }
 
     function testBurn() public {
         vm.prank(alice);
         erc721.safeMint(bob);
-        
+
         vm.prank(alice);
-        vm.expectRevert(); // "Only token owner can burn the token"        
+        vm.expectRevert(); // "Only token owner can burn the token"
         erc721.burn(0);
 
         vm.prank(bob);
-        vm.expectCall(
-            address(erc721),
-            abi.encodeCall(erc721.burn, 0)
-        );
+        vm.expectCall(address(erc721), abi.encodeCall(erc721.burn, 0));
         erc721.burn(0);
-        uint bob_balance = erc721.balanceOf(bob);
+        uint256 bob_balance = erc721.balanceOf(bob);
         assertEq(bob_balance, 0);
 
         vm.expectRevert(); // "Token does not exist"
@@ -158,10 +149,7 @@ contract TestERC721 is Test {
         vm.prank(bob);
         erc721.pause();
 
-        vm.expectCall(
-            address(erc721),
-            abi.encodeCall(erc721.pause, ())
-        );
+        vm.expectCall(address(erc721), abi.encodeCall(erc721.pause, ()));
         vm.prank(alice);
         erc721.pause();
         bool paused = erc721.paused();
@@ -173,10 +161,7 @@ contract TestERC721 is Test {
         vm.prank(bob);
         erc721.grantRole(keccak256("PAUSER_ROLE"), alice);
 
-        vm.expectCall(
-            address(erc721),
-            abi.encodeCall(erc721.grantRole, (keccak256("PAUSER_ROLE"), bob))
-        );
+        vm.expectCall(address(erc721), abi.encodeCall(erc721.grantRole, (keccak256("PAUSER_ROLE"), bob)));
         vm.prank(alice);
         erc721.grantRole(keccak256("PAUSER_ROLE"), bob);
         bool hasRole = erc721.hasRole(keccak256("PAUSER_ROLE"), bob);
